add auth routes, comment routes, authentication middleware, redis server

+++ controllers/userController.mjs
@@ -18,7 +18,7 @@ const createUser = async (req, res) => {
     }
 
     try {
-        const uuid = await userService.createUser({ email, password, name, username, presentation });
+        const uuid = await userService.createUser(email, password, name, username, presentation);
         res.status(201).json({ data: uuid, message: 'User created successfully' });
     } catch (error) {
         res.status(400).json({ error: error.message });
@@ -29,7 +29,23 @@ const findUserByUUID = async (req, res) => {
     const { uuid } = req.params;
     try {
         const user = await userService.findUserByUUID(uuid);
-        res.status(200).json(user);
+
+        if (req?.user?.is_admin != true) {
+            return res.status(200).json(user);
+        }
+        res.status(200).json(
+            users.map(user => {
+                return {
+                    user: {
+                        uuid: user.user.uuid,
+                        username: user.user.username,
+                        presentation: user.user.presentation,
+                        is_admin: user.user.is_admin
+                    },
+                    posts: user.posts
+                };
+            })
+        );
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
@@ -40,7 +56,7 @@ const findUserPostsByUUID = async (req, res) => {
     const { uuid } = req.params;
     const { page = 1, limit = 25 } = req.query;
     try {
-        const posts = await postService.findPostsByUserUUID(uuid, page, limit > 25 || limit < 1  ? 25 : limit); // min limit: 1, max limit: 25
+        const posts = await postService.findPostsByUserUUID(uuid, page, limit > 25 || limit < 1 ? 25 : limit); // min limit: 1, max limit: 25
         res.status(200).json(posts);
     } catch (error) {
         res.status(500).json({ error: error.message });
@@ -51,7 +67,18 @@ const findUsers = async (req, res) => {
     const { page = 1, limit = 25 } = req.query;
     try {
         const users = await userService.findUsers(page, limit > 25 || limit < 1 ? 25 : limit); // min limit: 1, max limit: 25
-        res.status(200).json(users);
+        if (req?.user?.is_admin != true) {
+            return res.status(200).json(users);
+        }
+        res.status(200).json({
+            user: {
+                uuid: user.user.uuid,
+                username: user.user.username,
+                presentation: user.user.presentation,
+                is_admin: user.user.is_admin
+            },
+            posts: user.posts
+        });
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
@@ -59,6 +86,11 @@ const findUsers = async (req, res) => {
 
 const removeUserByUUID = async (req, res) => {
     const { uuid } = req.params;
+
+    if (req.user.uuid != uuid && req.user.is_admin) {
+        return res.status(403).json({ message: "Insufficient permissions" });
+    }
+
     try {
         await userService.removeUserByUUID(uuid);
         res.status(200).json({ message: 'User deleted successfully' });
@@ -69,7 +101,11 @@ const removeUserByUUID = async (req, res) => {
 
 const updateUserByUUID = async (req, res) => {
     const { uuid } = req.params;
-    const { email, name, username, presentation, isAdmin } = req.body;
+    const { email, name, username, presentation, is_admin } = req.body;
+
+    if (req.user.uuid != uuid && req.user.is_admin) {
+        return res.status(403).json({ message: "Insufficient permissions" });
+    }
 
     if (!validateEmail(email)) {
         return res.status(400).json({ error: 'Invalid email format' });
@@ -84,7 +120,7 @@ const updateUserByUUID = async (req, res) => {
     }
 
     try {
-        await userService.updateUserByUUID(uuid, { email, name, username, presentation, isAdmin });
+        await userService.updateUserByUUID(uuid, { email, name, username, presentation, is_admin });
         res.status(200).json({ message: 'User updated successfully' });
     } catch (error) {
         res.status(500).json({ error: error.message });
@@ -94,6 +130,10 @@ const updateUserByUUID = async (req, res) => {
 const updateUserPasswordByUUID = async (req, res) => {
     const { uuid } = req.params;
     const { password } = req.body;
+
+    if (req.user.uuid != uuid && req.user.is_admin) {
+        return res.status(403).json({ message: "Insufficient permissions" });
+    }
 
     if (!password) {
         return res.status(400).json({ error: 'Password is required' });
